Move Recruiter Job DTO to DTO package
